requirements.txt
pandas==1.0.5
pandas_datareader==0.9.0
numpy==1.20.3
matplotlib==3.4.2
streamlit==0.82.0
Pillow==8.2.0
yfinance==0.1.63
